float

+++ Parameters.java
@@ -1,23 +1,23 @@
 
 public class Parameters {
 
-	public static float ALBEDO_WHITE = 0.75;
+	public static float ALBEDO_WHITE = (float)0.75;
 
-	public static float ALBEDO_BLACK = 0.25;
+	public static float ALBEDO_BLACK = (float)0.25;
 
-	public static float ALBEO_GROUND = 0.4;
+	public static float ALBEO_GROUND = (float)0.4;
 
-	public static float LUMINOSITY = 0.8;
+	public static float LUMINOSITY = (float)0.8;
 
-	public static float DIFFUSION_RATE = 0.5 ;
+	public static float DIFFUSION_RATE = (float)0.5 ;
 
 	public static int SENCERIO = 1;
 
 	public static int DAISY_LIFE_EXPECTANCY = 25;
 
-	public static float START_WHITE = 0.2;
+	public static float START_WHITE = (float)0.2;
 
-	public static float START_BLACK = 0.2 ;
+	public static float START_BLACK = (float)0.2 ;
 
 	public static int START_TEMP = 0;
